Put materials on a context different from context_default
This is needed, as otherwise whenever xcoll is imported, the
materials are allocated on the default context. This is not in
itself a problem, but causes a problem for the test run of xtrack,
where emptiness of the default context is checked (as a means of
detecting memory leaks -- this was a problem in the past).
Rather than disabling the check, we can allocate the materials on
a different context, as a means of specifying that these are valid
allocations. Maybe a better solution can be found, but for now
this seems the best.

Signed-off-by: Szymon Łopaciuk <szymon@lopaciuk.eu>